+++ themes-default/slim/src/components/manual-post-process.vue
@@ -8,7 +8,7 @@
 
                 <div class="col-xs-12 col-lg-9">
                     <config-template label="Process Method to be used" label-for="process_method">
-                        <select id="process_method" name="process_method" v-model="processMethod" class="form-control input-sm">
+                        <select id="process_method" name="process_method" :value="postprocessing.processMethod" @update="processMethod = $event" class="form-control input-sm">
                             <option v-for="option in availableMethods" :value="option.value" :key="option.value">
                                 {{option.text}}
                             </option>
@@ -68,6 +68,10 @@ export default {
         FileBrowser,
         ConfigTemplate,
         ConfigToggleSlider
+    },
+    mounted() {
+        this.processMethod = this.postprocessing.processMethod;
+        this.path = this.postprocessing.showDownloadDir;
     },
     data() {
         return {